BACKEND-WEB-COLLECTOR/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=1.10,<3
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
readability-lxml>=0.8.1
yake>=0.4.8
playwright>=1.44
prometheus-client>=0.20
sentry-sdk>=2.0
kafka-python>=2.0
python-dotenv>=1.0
